Only 64-bit right before pointers

+++ cut_cross_entropy/indexed_dot.py
@@ -43,15 +43,17 @@ def _indexed_neg_dot_forward_kernel(
     group_id = pid // num_d_in_group
     first_pid_b = group_id * GROUP_B
     group_size_b = min(num_b_chunks - first_pid_b, GROUP_B)
-    pid_b = (first_pid_b + ((pid % num_d_in_group) % group_size_b)).to(tl.int64)
-    pid_d = ((pid % num_d_in_group) // group_size_b).to(tl.int64)
+    pid_b = first_pid_b + ((pid % num_d_in_group) % group_size_b)
+    pid_d = (pid % num_d_in_group) // group_size_b
 
-    offs_b = (tl.arange(0, BLOCK_B) + pid_b * BLOCK_B).to(tl.int64)
+    offs_b = tl.arange(0, BLOCK_B) + pid_b * BLOCK_B
     if HAS_VALIDS:
-        offs_b = tl.load(Valids + stride_vb * offs_b, mask=offs_b < B, other=BMax).to(tl.int64)
+        offs_b = tl.load(Valids + stride_vb * offs_b.to(tl.int64), mask=offs_b < B, other=BMax)
 
-    offs_d = (tl.arange(0, BLOCK_D) + pid_d * BLOCK_D).to(tl.int64)
-    e_ptrs = E + (stride_eb * offs_b[:, None] + stride_ed * offs_d[None, :])
+    offs_d = tl.arange(0, BLOCK_D) + pid_d * BLOCK_D
+    e_ptrs = E + (
+        stride_eb * offs_b[:, None].to(tl.int64) + stride_ed * offs_d[None, :].to(tl.int64)
+    )
 
     e_mask = offs_b[:, None] < BMax
     if not EVEN_D:
@@ -62,9 +64,9 @@ def _indexed_neg_dot_forward_kernel(
     if HAS_SHIFT:
         offs_b = offs_b + shift
 
-    inds = tl.load(Inds + stride_ib * offs_b, mask=offs_b < BMax, other=V)
+    inds = tl.load(Inds + stride_ib * offs_b.to(tl.int64), mask=offs_b < BMax, other=V)
 
-    c_ptrs = C + (inds[:, None] * stride_cv + offs_d[None, :] * stride_cd)
+    c_ptrs = C + (inds[:, None].to(tl.int64) * stride_cv + offs_d[None, :].to(tl.int64) * stride_cd)
 
     c_mask = inds[:, None] < V
     if not EVEN_D:
@@ -80,7 +82,7 @@ def _indexed_neg_dot_forward_kernel(
         bias = bias.to(tl.float32)
         neg_dot -= bias
 
-    offs_b = (tl.arange(0, BLOCK_B) + pid_b * BLOCK_B).to(tl.int64)
+    offs_b = tl.arange(0, BLOCK_B) + pid_b * BLOCK_B
     out_ptrs = Out + offs_b
     tl.atomic_add(out_ptrs, neg_dot.to(out_ptrs.dtype.element_ty), mask=offs_b < B)
